requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
requests==2.28.1
flask==2.0.1
flask-socketio==5.1.1
flask-cors==3.0.10
python-telegram-bot==13.7
werkzeug==2.0.1
eventlet==0.33.0
aiohttp>=3.8.0